Refactor: Move CourseCompletion and extraction logic to ProgressRepository
- Moved `CourseCompletion` data class to `model` package.
- Abstracted UI-level logic for resolving completed courses from `BellDashboardViewModel` into `ProgressRepository`.
- Used `dagger.Lazy` in `ProgressRepositoryImpl` to prevent circular dependency with `CoursesRepository`.
