+++ app/src/main/java/org/ole/planet/myplanet/model/CourseCompletion.kt
@@ -1,0 +1,3 @@
+package org.ole.planet.myplanet.model
+
+data class CourseCompletion(val courseId: String?, val courseTitle: String?)
