+++ app/src/main/java/org/ole/planet/myplanet/repository/ProgressRepository.kt
@@ -2,6 +2,7 @@ package org.ole.planet.myplanet.repository
 
 import com.google.gson.JsonArray
 import com.google.gson.JsonObject
+import org.ole.planet.myplanet.model.CourseCompletion
 import org.ole.planet.myplanet.model.RealmCourseProgress
 import org.ole.planet.myplanet.model.RealmCourseStep
 
@@ -10,6 +11,7 @@ interface ProgressRepository {
     suspend fun getCurrentProgress(steps: List<RealmCourseStep?>?, userId: String?, courseId: String?): Int
     suspend fun fetchCourseData(userId: String?): JsonArray
     suspend fun getProgressRecords(userId: String?): List<RealmCourseProgress>
+    suspend fun getCompletedCourses(userId: String): List<CourseCompletion>
     suspend fun saveCourseProgress(
         userId: String?,
         planetCode: String?,
