+++ app/src/main/java/org/ole/planet/myplanet/repository/ProgressRepositoryImpl.kt
@@ -10,6 +10,7 @@ import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
 import org.ole.planet.myplanet.data.DatabaseService
 import org.ole.planet.myplanet.di.RealmDispatcher
+import org.ole.planet.myplanet.model.CourseCompletion
 import org.ole.planet.myplanet.model.RealmAnswer
 import org.ole.planet.myplanet.model.RealmCourseProgress
 import org.ole.planet.myplanet.model.RealmCourseStep
@@ -24,7 +25,8 @@ import org.ole.planet.myplanet.utils.JsonUtils
 class ProgressRepositoryImpl @Inject constructor(
     databaseService: DatabaseService,
     @RealmDispatcher realmDispatcher: CoroutineDispatcher,
-    private val dispatcherProvider: DispatcherProvider
+    private val dispatcherProvider: DispatcherProvider,
+    private val coursesRepositoryLazy: dagger.Lazy<CoursesRepository>
 ) : RealmRepository(databaseService, realmDispatcher), ProgressRepository {
     override suspend fun getCourseProgress(userId: String?): HashMap<String?, JsonObject> = withContext(dispatcherProvider.io) {
         val mycourses = queryList(RealmMyCourse::class.java) {
@@ -169,6 +171,37 @@ class ProgressRepositoryImpl @Inject constructor(
         queryList(RealmCourseProgress::class.java) {
             equalTo("userId", userId)
         }
+    }
+
+    override suspend fun getCompletedCourses(userId: String): List<CourseCompletion> = withContext(dispatcherProvider.io) {
+        val myCourses = coursesRepositoryLazy.get().getMyCourses(userId)
+        val allProgressRecords = getProgressRecords(userId)
+
+        val completedCourses = mutableListOf<CourseCompletion>()
+        myCourses.forEachIndexed { index, course ->
+            val hasValidId = !course.courseId.isNullOrBlank()
+            val hasValidTitle = !course.courseTitle.isNullOrBlank()
+
+            // Get progress records for this specific course
+            val courseProgressRecords = allProgressRecords.filter { it.courseId == course.courseId }
+
+            // Count UNIQUE steps that are passed (matches web: step.passed === true)
+            val passedStepNumbers = courseProgressRecords
+                .filter { it.passed }
+                .map { it.stepNum }
+                .toSet()
+            val passedSteps = passedStepNumbers.size
+            val totalSteps = course.courseSteps?.size ?: 0
+
+            // Web logic: ALL steps must be passed AND course must have at least one step
+            val allStepsPassed = passedSteps == totalSteps && totalSteps > 0
+
+            // Match web behavior: Show badge if ALL steps are passed AND course has steps
+            if (allStepsPassed && hasValidId && hasValidTitle) {
+                completedCourses.add(CourseCompletion(course.courseId, course.courseTitle))
+            }
+        }
+        completedCourses
     }
 
     override suspend fun saveCourseProgress(
